Refactor strategy runner to group builds
Because Travis' jobs can execute concurrently, none of the jobs
is guaranteed to see the knowledge of priorly executed jobs of
the same build.
Therefore the runner adopts a turn-based model, in which all
jobs of a build are first reprioritized, and then the strategy
gets hold of this batch of results.
For this to work, each strategy has to split reprioritization
of a single build and incoporating build outcome.